+++ src/refactoring/components/admin/index.ts
@@ -1,3 +1,5 @@
 export * from "./AdminPage";
-export * from "./NewProductCard";
 export * from "./AdminProductCard";
+export * from "./CouponSection";
+export * from "./NewProductCard";
+export * from "./ProductSection";
